requirements.txt
opencv-python
face-recognition
numpy
requests
playsound==1.2.2
python-dotenv
twilio